make plugins in CalibTracker/SiPixelESProducers/test self-contained
